byml: Add a Writer
This adds a Writer class that can be used to convert arbitrary dicts
and lists into BYAML.

A utility script is provided to convert YAML files to BYAML.

Tested with Breath of the Wild's LevelSensor.byml (both little and big
endian). File contents are correctly converted between the
human-readable and binary forms.

Hash node order in the binary form couldn't be preserved though, sadly,
because it seems that Nintendo's conversion tool removes ordering info
by sorting entries for hash node children but the actual children nodes
are still saved in the original order.