scripts: Handle yaz0 files in a nicer way
byml_to_yml: Automatically recognise and decompress yaz0 files.

yml_to_byml: If the target file extension starts with 's', which
indicates the file is yaz0 compressed, then automatically compress
the resulting byml with yaz0.

This relies on wszst and assumes that it can be found in the user path.